🚚 Move: some component files for home to `$lib/components/home/`
