+++ src/lib/components/home/GearsAndSettings.svelte
