+++ src/lib/components/home/MainVisual.svelte
