+++ src/lib/components/home/Sponsors.svelte
