+++ src/lib/components/home/sponsor_banner/SponsorBanner.svelte
